src/nyc_taxi_data_regression/pipelines/train/requirements.txt
kedro[pandas.CSVDataSet]~=0.16
kedro-mlflow~=0.4
scikit-learn>=0.22.1
